requirements.txt
# ==========================================
# RAG_APP Requirements
# FastAPI + Ollama + LangChain + PostgreSQL
# ==========================================

# ====================
# Web Framework & Server
# ====================
fastapi==0.109.1
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ====================
# Data Validation & Serialization
# ====================
pydantic==2.5.3
pydantic[email]==2.5.3

# ====================
# Database & ORM
# ====================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# ====================
# Authentication & Security
# ====================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ====================
# Environment Management
# ====================
python-dotenv==1.0.0

# ====================
# HTTP & Requests
# ====================
requests==2.31.0
httpx==0.25.2

# ====================
# Redis Caching
# ====================
redis==5.0.1

# ====================
# LangChain & LLM Integration
# ====================
langchain==0.1.20
langchain-community==0.0.38
langchain_text_splitters==0.0.1

# ====================
# LLM - Ollama Integration
# ====================
# Note: Ollama should be running separately in Docker
# ollama is accessed via HTTP at http://ollama:11434

# ====================
# Document Processing
# ====================
pypdf==4.0.1
pdf2image==1.16.3
pytesseract==0.3.10

# ====================
# Vector Search & Embeddings
# ====================
faiss-cpu==1.7.4
sentence-transformers==2.6.1
scikit-learn==1.3.2

# ====================
# Machine Learning Framework
# ====================
torch==2.1.2 --index-url https://download.pytorch.org/whl/cpu
transformers==4.36.2

# ====================
# Utilities
# ====================
python-slugify==8.0.1
python-dateutil==2.8.2
qrcode==8.0
Pillow==11.0.0

# ====================
# Development & Testing (Optional)
# ====================
# pytest==7.4.3
# pytest-asyncio==0.23.2
# black==23.12.1
# flake8==6.1.0
# isort==5.13.2

# ====================
# Logging & Monitoring (Optional)
# ====================
# python-json-logger==2.0.7
# sentry-sdk==1.38.0
huggingface-hub==0.20.3
pandas
pyodbc